rapid_layout/utils/__init__.py
# -*- encoding: utf-8 -*-
# @Author: SWHL
# @Contact: liekkaskono@163.com
import yaml

from .download_model import DownloadModel
from .infer_engine import OrtInferSession
from .load_image import LoadImage
from .logger import get_logger
from .post_prepross import PicoDetPostProcess
from .pre_procss import create_operators, transform
from .vis_res import VisLayout


def read_yaml(yaml_path):
    with open(yaml_path, "rb") as f:
        data = yaml.load(f, Loader=yaml.Loader)
    return data
